Seek output to same value as input
The FFmpeg wiki says:

> As of FFmpeg 2.1, combined seeking is still possible but I have yet to
> find a valid use case for it

Here is a valid use case for ya: when combining input seeking with -copyts,
the resulting output is almost unplayable with both mpv and TikTok (and
probably lots of other players, too). These players don't discard the time
between 00:00:00 and the -ss timestamp, so you have to stare at a black
screen like an idiot until playback finally begins.

mpv is at least smart enough to do the initial seeking for you (TikTok
isn't) but the timeline, seek bar, track slider or whatever the damn thing
is called is still broken. Letting FFmpeg do the output seeking solves this
problem, but without ALSO specifying the input seeking the entire video will
be transcoded and discarded except for the selected cut, which is a waste of
time and resources.

So, if anyone from @FFmpeg reads this, please tell me that I'm wrong or
update your Wiki. I still love you.