Lose mind over math
Ensure that the aspect ratio of the output video is never wider than that of
the input video. In other words, with a 4/3 input video, a ratio of 9/16
will still produce TikTok crop, but a ratio of 16/9 will produce a 4/3 crop.

Trust me, this makes sense.

Also, fix a bug in handling filenames with spaces.