Don't trust me
The previous commit was incorrect, the output was scaled incorrectly when
the aspect ratio was too wide. However, it was cropped correctly, and
because the crop ratio is what is displayed by `ffplay`, rather than the
actual size of the video, I didn't notice.

The math is now correct and I will never touch it again.